refactor: replace clap with rigging for CLI parsing and update dependencies
- Migrate CLI command parsing in kernel and system crates from clap to rigging::Commands.
- Group server management commands under the 'server' namespace ('server start', 'stop', 'restart', 'status', 'serve').
- Add '--uid' and '--new' flag options to the interactive 'chat' command.
- Update workspace dependencies: 'atoman' (0.7.1), 'rigging' (0.3.0), and introduce 'atoman-config'.
- Simplify agent skill routing: remove 'agent' field from 'TaskAction' and match skills directly via 'HashMap'.
- Rename 'javascript_eval' tool to 'js_eval'.
- Use '#[atoman::config]' macro for configuration management in 'Settings'.
- Bump package versions: 'osy' to 0.20.0 and 'osy-system' to 0.5.0.

+++ crates/share/src/session_id.rs
@@ -8,14 +8,14 @@ use std::{num::ParseIntError, str::FromStr};
 #[derive(Default, Debug, Display, Copy, Clone, Eq, PartialEq, Hash)]
 #[display(fmt = "{user_id}-{timestamp}-{salt}")]
 pub struct SessionId {
-    pub user_id: u128,
+    pub user_id: u64,
     pub timestamp: u128,
     pub salt: u16,
 }
 
 impl SessionId {
     /// Creates a new session ID
-    pub fn new(user_id: u128) -> Self {
+    pub fn new(user_id: u64) -> Self {
         let timestamp = std::time::SystemTime::now()
             .duration_since(std::time::UNIX_EPOCH)
             .expect("Time went backwards")
@@ -60,7 +60,7 @@ impl FromStr for SessionId {
         }
 
         let user_id = user_id_str
-            .parse::<u128>()
+            .parse::<u64>()
             .map_err(SessionIdError::InvalidUserId)?;
         let timestamp = timestamp_str
             .parse::<u128>()
